Add support for external links in header, add Anvil to Projects
Anvil is what is used to generate this static site, and is hosted
on GitHub. This commit also adds support for linking to external
webpages.

The presence of "http" in a link indicates it is an external link,
thus introducting the restriction that no local .yaml file can
have "http" in its filename.